LayoutTests/http/tests/websocket/connection-refusal-in-frame-resource-load-statistics.html

<!DOCTYPE html>
<head>
    <script src="../resources/js-test-pre.js"></script>
    <script src="/resourceLoadStatistics/resources/util.js"></script>
</head>
<body onload="runTest()">
<script>
    description('Construct a cross-site WebSocket in a frame with server-side refusal. The test passes if Resource Load Statistics logs it properly.');

    window.jsTestIsAsync = true;

    function completeTest() {
        if (testRunner.isStatisticsRegisteredAsSubresourceUnder("http://localhost", "http://127.0.0.1"))
            testPassed("localhost registered as subresource under 127.0.0.1.");
        else
            testFailed("localhost not registered as subresource under 127.0.0.1.");
        setEnableFeature(false, finishJSTest);
    }

    function runTest() {
        setEnableFeature(true, function() {
            if (testRunner.isStatisticsRegisteredAsSubresourceUnder("http://localhost", "http://127.0.0.1"))
                testFailed("localhost already registered as subresource under 127.0.0.1.");

            testRunner.setStatisticsNotifyPagesWhenDataRecordsWereScanned(true);
            testRunner.installStatisticsDidScanDataRecordsCallback(completeTest);

            let iframeElement = document.createElement("iframe");
            iframeElement.src = "resources/localhost-websocket-connect.html";
            iframeElement.id = "testIframe";
            document.body.appendChild(iframeElement);
        });
    }
</script>
<script src="../resources/js-test-post.js"></script>
</body>
