LayoutTests/platform/glib/http/tests/websocket/connection-refusal-in-frame-resource-load-statistics-expected.txt

CONSOLE MESSAGE: WebSocket connection to 'ws://localhost/' failed: The server did not accept the WebSocket handshake.
Construct a cross-site WebSocket in a frame with server-side refusal. The test passes if Resource Load Statistics logs it properly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS localhost registered as subresource under 127.0.0.1.
PASS successfullyParsed is true

TEST COMPLETE
